# openapi/jobtopbob.yaml
openapi: 3.1.0
info:
  title: JobTopBob API
  version: 0.0.2
  description: JobTopBob API server
servers:
  - url: http://localhost:8080
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by Better Auth, validated via JWKS
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    PaginatedResponse:
      type: object
      required: [data, total, page, per_page]
      properties:
        data:
          type: array
          items: {}
        total:
          type: integer
          format: int64
        page:
          type: integer
        per_page:
          type: integer
    Job:
      type: object
      required: [id, user_id, title, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
          nullable: true
        stage_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        status:
          type: string
          nullable: true
        close_reason:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        source_url:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        location_type:
          type: string
          nullable: true
        salary_min:
          type: integer
          nullable: true
        salary_max:
          type: integer
          nullable: true
        salary_market:
          type: integer
          nullable: true
        salary_offered:
          type: integer
          nullable: true
        salary_currency:
          type: string
          nullable: true
        interest:
          type: integer
          nullable: true
        suitability:
          type: integer
          nullable: true
        suitability_reason:
          type: string
          nullable: true
        resume_version_id:
          type: string
          format: uuid
          nullable: true
        jd_raw:
          type: string
          nullable: true
        jd_snapshot:
          type: object
          nullable: true
        applied_at:
          type: string
          format: date-time
          nullable: true
        follow_up_at:
          type: string
          format: date-time
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        job_type:
          type: string
          nullable: true
        job_level:
          type: string
          nullable: true
        salary_interval:
          type: string
          nullable: true
        application_url:
          type: string
          nullable: true
        experience_range:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
          nullable: true
        closed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        company_name:
          type: string
          nullable: true
          description: Joined from companies table
        company_logo_url:
          type: string
          nullable: true
          description: Company logo URL from storage
        stage_name:
          type: string
          nullable: true
          description: Joined from stages table
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
...
    Tag:
      type: object
      required: [id, user_id, name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreateJobRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        company_id:
          type: string
          format: uuid
        stage_id:
          type: string
          format: uuid
        status:
          type: string
        source:
          type: string
        source_url:
          type: string
        location:
          type: string
        location_type:
          type: string
        salary_min:
          type: integer
        salary_max:
          type: integer
        salary_offered:
          type: integer
        salary_currency:
          type: string
        salary_interval:
          type: string
        interest:
          type: integer
        jd_raw:
          type: string
        applied_at:
          type: string
          format: date-time
        follow_up_at:
          type: string
          format: date-time
        deadline:
          type: string
          format: date-time
        job_type:
          type: string
        job_level:
          type: string
        application_url:
          type: string
        experience_range:
          type: string
    UpdateJobRequest:
      type: object
      properties:
        title:
          type: string
        company_id:
          type: string
          format: uuid
          nullable: true
        stage_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          nullable: true
        close_reason:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        source_url:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        location_type:
          type: string
          nullable: true
        salary_min:
          type: integer
          nullable: true
        salary_max:
          type: integer
          nullable: true
        salary_market:
          type: integer
          nullable: true
        salary_offered:
          type: integer
          nullable: true
        salary_currency:
          type: string
          nullable: true
        interest:
          type: integer
          nullable: true
        suitability:
          type: integer
          nullable: true
        suitability_reason:
          type: string
          nullable: true
        resume_version_id:
          type: string
          format: uuid
          nullable: true
        jd_raw:
          type: string
          nullable: true
        applied_at:
          type: string
          format: date-time
          nullable: true
        follow_up_at:
          type: string
          format: date-time
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        job_type:
          type: string
          nullable: true
        job_level:
          type: string
          nullable: true
        salary_interval:
          type: string
          nullable: true
        application_url:
          type: string
          nullable: true
        experience_range:
          type: string
          nullable: true
        closed_at:
          type: string
          format: date-time
          nullable: true
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/jobs:
    get:
      summary: List jobs with filters and pagination
      parameters:
        - name: statuses
          in: query
          description: Comma-separated list of statuses to filter by
          schema:
            type: string
        - name: stage_ids
          in: query
          description: Comma-separated list of stage UUIDs to filter by
          schema:
            type: string
        - name: location_types
          in: query
          description: Comma-separated list of location types to filter by
          schema:
            type: string
        - name: sources
          in: query
          description: Comma-separated list of sources to filter by
          schema:
            type: string
        - name: job_types
          in: query
          description: Comma-separated list of job types to filter by
          schema:
            type: string
        - name: job_levels
          in: query
          description: Comma-separated list of job levels to filter by
          schema:
            type: string
        - name: tag_ids
          in: query
          description: Comma-separated list of tag UUIDs to filter by
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
          description: Search in title, company name, location
        - name: created_after
          in: query
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 25
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, title, applied_at, salary_min, deadline]
            default: created_at
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated list of jobs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Job"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create a new job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/jobs/{id}:
    get:
      summary: Get a job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobRequest"
      responses:
        "200":
          description: Job updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Job deleted
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...